// api/src/types/interface.d.ts
interface Field {
    faculty?: number,
    fieldCode?: string;
    fieldNameKh?: string;
    fieldNameEn?: string;
    changedByUser?: string;
    clientIp?: string;
    sessionInfo?: string;
    status?: 'Active' | 'Inactive';
}
// ...
export interface SearchParams{
    keyword?: string;
    page?: number;
    limit?: number;
}